moved about di classes
