+++ app/src/main/java/com/freshdigitable/udonroad2/di/AppComponent.kt
@@ -21,9 +21,9 @@ import com.freshdigitable.udonroad2.AppApplication
 import com.freshdigitable.udonroad2.AppFileProviderImpl
 import com.freshdigitable.udonroad2.AppSetup
 import com.freshdigitable.udonroad2.AppSetupModule
-import com.freshdigitable.udonroad2.data.db.DatabaseModule
 import com.freshdigitable.udonroad2.data.impl.ListOwnerRepository
 import com.freshdigitable.udonroad2.data.impl.di.RepositoryModule
+import com.freshdigitable.udonroad2.data.local.di.DatabaseModule
 import com.freshdigitable.udonroad2.data.restclient.AppTwitterModule
 import com.freshdigitable.udonroad2.data.restclient.TwitterModule
 import com.freshdigitable.udonroad2.input.di.MediaChooserModule
