+++ data/repository/src/main/java/com/freshdigitable/udonroad2/data/impl/di/ListRepositoryModule.kt
@@ -18,16 +18,16 @@ package com.freshdigitable.udonroad2.data.impl.di
 
 import com.freshdigitable.udonroad2.data.ListRepository
 import com.freshdigitable.udonroad2.data.PagedListProvider
-import com.freshdigitable.udonroad2.data.db.LocalListDataSourceModule
 import com.freshdigitable.udonroad2.data.db.LocalListDataSourceProvider
-import com.freshdigitable.udonroad2.data.db.PagedListDataSourceFactoryModule
 import com.freshdigitable.udonroad2.data.db.PagedListDataSourceFactoryProvider
 import com.freshdigitable.udonroad2.data.impl.ListRepositoryImpl
 import com.freshdigitable.udonroad2.data.impl.PagedListProviderImpl
-import com.freshdigitable.udonroad2.data.restclient.CustomTimelineDataSourceModule
+import com.freshdigitable.udonroad2.data.local.di.LocalListDataSourceModule
+import com.freshdigitable.udonroad2.data.local.di.PagedListDataSourceFactoryModule
 import com.freshdigitable.udonroad2.data.restclient.RemoteListDataSourceProvider
-import com.freshdigitable.udonroad2.data.restclient.TweetTimelineDataSourceModule
-import com.freshdigitable.udonroad2.data.restclient.UserListDataSourceModule
+import com.freshdigitable.udonroad2.data.restclient.di.CustomTimelineDataSourceModule
+import com.freshdigitable.udonroad2.data.restclient.di.TweetTimelineDataSourceModule
+import com.freshdigitable.udonroad2.data.restclient.di.UserListDataSourceModule
 import com.freshdigitable.udonroad2.model.QueryType
 import com.freshdigitable.udonroad2.model.app.AppExecutor
 import dagger.BindsInstance
